xml/parsing/manakai/ns-elements-1.dat

#data
<:abc/>
#errors
1;1;no XML decl;;;s
#document
| <{} :abc>

#data
<:abc xmlns:="a"/>
#errors
1;1;no XML decl;;;s
#document
| <{} :abc>
|   xmlns:="a"

#data
<:abc xmlns="a"/>
#errors
1;1;no XML decl;;;s
#document
| <{} :abc>
|   xmlns xmlns="a"

#data
<a: xmlns="a" xmlns:a="b"/>
#errors
1;1;no XML decl;;;s
#document
| <{} a:>
|   xmlns a="b"
|   xmlns xmlns="a"

#data
<a:/>
#errors
1;1;no XML decl;;;s
#document
| <{} a:>

#data
<a:b xmlns="c"/>
#errors
1;1;no XML decl;;;s
#document
| <{} a:b>
|   xmlns xmlns="c"

#data
<: xmlns="a"/>
#errors
1;1;no XML decl;;;s
#document
| <{} :>
|   xmlns xmlns="a"

#data
<: xmlns:="a"/>
#errors
1;1;no XML decl;;;s
#document
| <{} :>
|   xmlns:="a"

#data
<z><:abc/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} :abc>

#data
<z><:abc xmlns:="a"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} :abc>
|     xmlns:="a"

#data
<z><:abc xmlns="a"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} :abc>
|     xmlns xmlns="a"

#data
<z><a: xmlns="a" xmlns:a="b"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} a:>
|     xmlns a="b"
|     xmlns xmlns="a"

#data
<z><a:/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} a:>

#data
<z><a:b xmlns="c"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} a:b>
|     xmlns xmlns="c"

#data
<z><: xmlns="a"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} :>
|     xmlns xmlns="a"

#data
<z><: xmlns:="a"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} :>
|     xmlns:="a"

#data
<a:b:c/>
#errors
1;1;no XML decl;;;s
#document
| <{} a:b:c>

#data
<a:b:c xmlns:a="x"/>
#errors
1;1;no XML decl;;;s
#document
| <{x} b:c>
|   xmlns a="x"

#data
<a:b:c xmlns:a:b="x"/>
#errors
1;1;no XML decl;;;s
#document
| <{} a:b:c>
|   xmlns a:b="x"

#data
<::a xmlns::="x"/>
#errors
1;1;no XML decl;;;s
#document
| <{} ::a>
|   xmlns :="x"

#data
<z><a:b:c/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} a:b:c>

#data
<z><a:b:c xmlns:a="x"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{x} b:c>
|     xmlns a="x"

#data
<z><a:b:c xmlns:a:b="x"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} a:b:c>
|     xmlns a:b="x"

#data
<z><::a xmlns::="x"/></z>
#errors
1;1;no XML decl;;;s
#document
| <{} z>
|   <{} ::a>
|     xmlns :="x"

#data
<a:b xmlns:a="http://www.w3.org/1999/xhtml"><a:template><p>hoge</p></a:template></a:b>
#errors
1;1;no XML decl;;;s
#document
| <b>
|   xmlns a="http://www.w3.org/1999/xhtml"
|   <template>
|     content
|       <{} p>
|         "hoge"

#data
<a:b xmlns:a="http://www.w3.org/1999/xhtml"><a:template><p>hoge</p></a:template></b>
#errors
1;1;no XML decl;;;s
1;81;stray end tag;;b;m
1;85;in body:#eof;b;;m
#document
| <b>
|   xmlns a="http://www.w3.org/1999/xhtml"
|   <template>
|     content
|       <{} p>
|         "hoge"
